This commit was generated by cvs2svn to compensate for changes in r13049, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@13050 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/c64io.c
@@ -94,7 +94,7 @@ BYTE REGPARM1 io2_read(WORD addr)
     if (mem_cartridge_type != CARTRIDGE_NONE)
         return cartridge_read_io2(addr);
     if (emu_id_enabled && addr >= 0xdfa0)
-        return emuid_read(addr - 0xdfa0);
+        return emuid_read((WORD)(addr - 0xdfa0));
     if (reu_enabled)
         return reu_read((WORD)(addr & 0x0f));
